refactor: seperate module directory by Model, Logic(controller), View
