+++ src/Logic/validator.js
