+++ src/Model/vending-machine.js
