+++ src/View/render.js
@@ -1,0 +1,3 @@
+export const render = (instance) => {
+    
+}
