+++ src/index.js
@@ -1,9 +1,9 @@
 import createHeaderPage from './components/header-page.js';
-import Controller from './machine/controller.js';
+import Controller from './Logic/controller.js';
 import { checkStorage, loadStorage, updateStorage } from './utils/storage.js';
 import createEventListeners from './components/eventListeners.js';
-import VendingMachine from './machine/vending-machine.js'
-
+import VendingMachine from './Model/vending-machine.js'
+import { render } from './View/render.js'
 createHeaderPage();
 
 const vendingMachine = new VendingMachine();
@@ -12,6 +12,8 @@ if(checkStorage()){
 }else {
     updateStorage(vendingMachine);
 };
+
+render(vendingMachine);
 
 const controller = new Controller();
 createEventListeners(controller);
